+++ src/main/activemq/connector/openwire/marshal/v2/PartialCommandMarshaller.cpp
@@ -64,7 +64,7 @@ int PartialCommandMarshaller::tightMarshal1( OpenWireFormat* wireFormat, DataStr
 
     int rc = BaseDataStreamMarshaller::tightMarshal1( wireFormat, dataStructure, bs );
     bs->writeBoolean( info->getData().size() != 0 );
-    rc += info->getData().size() == 0 ? 0 : info->getData().size() + 4;
+    rc += (int)(info->getData().size() == 0 ? 0 : info->getData().size() + 4);
 
     return rc + 4;
 }
@@ -78,7 +78,7 @@ void PartialCommandMarshaller::tightMarshal2( OpenWireFormat* wireFormat, DataSt
         dynamic_cast<PartialCommand*>( dataStructure );
     dataOut->write( info->getCommandId() );
     if( bs->readBoolean() ) {
-        dataOut->write( info->getData().size() );
+        dataOut->write( (int)info->getData().size() );
         dataOut->write( &info->getData()[0], info->getData().size() );
     }
 }
@@ -103,7 +103,7 @@ void PartialCommandMarshaller::looseMarshal( OpenWireFormat* wireFormat, DataStr
     dataOut->write( info->getCommandId() );
     dataOut->write( info->getData().size() != 0 );
     if( info->getData().size() != 0 ) {
-        dataOut->write( info->getData().size() );
+        dataOut->write( (int)info->getData().size() );
         dataOut->write( &info->getData()[0], info->getData().size() );
     }
 }
