+++ src/main/activemq/connector/openwire/marshal/v2/XATransactionIdMarshaller.cpp
@@ -65,9 +65,9 @@ int XATransactionIdMarshaller::tightMarshal1( OpenWireFormat* wireFormat, DataSt
 
     int rc = TransactionIdMarshaller::tightMarshal1( wireFormat, dataStructure, bs );
     bs->writeBoolean( info->getGlobalTransactionId().size() != 0 );
-    rc += info->getGlobalTransactionId().size() == 0 ? 0 : info->getGlobalTransactionId().size() + 4;
+    rc += (int)(info->getGlobalTransactionId().size() == 0 ? 0 : info->getGlobalTransactionId().size() + 4);
     bs->writeBoolean( info->getBranchQualifier().size() != 0 );
-    rc += info->getBranchQualifier().size() == 0 ? 0 : info->getBranchQualifier().size() + 4;
+    rc += (int)(info->getBranchQualifier().size() == 0 ? 0 : info->getBranchQualifier().size() + 4);
 
     return rc + 4;
 }
@@ -81,11 +81,11 @@ void XATransactionIdMarshaller::tightMarshal2( OpenWireFormat* wireFormat, DataS
         dynamic_cast<XATransactionId*>( dataStructure );
     dataOut->write( info->getFormatId() );
     if( bs->readBoolean() ) {
-        dataOut->write( info->getGlobalTransactionId().size() );
+        dataOut->write( (int)info->getGlobalTransactionId().size() );
         dataOut->write( &info->getGlobalTransactionId()[0], info->getGlobalTransactionId().size() );
     }
     if( bs->readBoolean() ) {
-        dataOut->write( info->getBranchQualifier().size() );
+        dataOut->write( (int)info->getBranchQualifier().size() );
         dataOut->write( &info->getBranchQualifier()[0], info->getBranchQualifier().size() );
     }
 }
@@ -111,12 +111,12 @@ void XATransactionIdMarshaller::looseMarshal( OpenWireFormat* wireFormat, DataSt
     dataOut->write( info->getFormatId() );
     dataOut->write( info->getGlobalTransactionId().size() != 0 );
     if( info->getGlobalTransactionId().size() != 0 ) {
-        dataOut->write( info->getGlobalTransactionId().size() );
+        dataOut->write( (int)info->getGlobalTransactionId().size() );
         dataOut->write( &info->getGlobalTransactionId()[0], info->getGlobalTransactionId().size() );
     }
     dataOut->write( info->getBranchQualifier().size() != 0 );
     if( info->getBranchQualifier().size() != 0 ) {
-        dataOut->write( info->getBranchQualifier().size() );
+        dataOut->write( (int)info->getBranchQualifier().size() );
         dataOut->write( &info->getBranchQualifier()[0], info->getBranchQualifier().size() );
     }
 }
